requirements.txt
fastapi
uvicorn
google-genai
youtube-transcript-api
python-dotenv